refactor: auth API 경로 변경 및 타입 정의 통합, 온보딩 페이지 성별 처리 방식 개선
