+++ omechu-app/src/app/(auth)/sign-in/components/SignInForm.tsx
@@ -12,9 +12,10 @@ import Checkbox from "@/auth/components/Checkbox";
 import SquareButton from "@/components/common/button/SquareButton";
 import Input from "@/components/common/Input";
 import Toast from "@/components/common/Toast";
-import { useLoginMutation } from "@/auth/hooks/useAuth";
+import { useAuthStore } from "@/auth/store";
 import { loginSchema, LoginFormValues } from "@/auth/schemas/auth.schema";
-import type { ApiResponse, LoginSuccessData } from "@/auth/api/auth";
+import { useLoginMutation } from "@/lib/hooks/useAuth";
+import type { ApiResponse, LoginSuccessData } from "@/lib/api/auth";
 
 export default function SignInForm() {
   const [showToast, setShowToast] = useState(false);
